added non-real functions and tests for them
