+++ lib/function/non-real/im.js
@@ -48,7 +48,7 @@ function factory (type, config, load, typed) {
     },
 
     'Quaternion': function (x) {
-      return x.i();
+      return x.i;
     }
   });
