+++ lib/function/non-real/index.js
@@ -2,5 +2,7 @@ module.exports = [
   require('./arg'),
   require('./conj'),
   require('./im'),
-  require('./re')
+  require('./re'),
+  require('./j'),
+  require('./k')
 ];
