+++ lib/function/non-real/j.js
@@ -30,7 +30,7 @@ function factory (type, config, load, typed) {
    *            A complex number or array with complex numbers
    * @return {number | BigNumber | Array | Matrix} The real part of x
    */
-  var re = typed('j', {
+  var j = typed('j', {
     'number | Complex': function (x) {
       return 0;
     },
@@ -40,18 +40,18 @@ function factory (type, config, load, typed) {
     },
 
     'Array | Matrix': function (x) {
-      return deepMap(x, re);
+      return deepMap(x, j);
     },
 
     'Quaternion': function (x) {
-      return x.j();
+      return x.j;
     }
   });
 
-  re.toTex = {1: '\\Re\\left\\lbrace${args[0]}\\right\\rbrace'};
+  j.toTex = {1: 'unfinished'};
 
-  return re;
+  return j;
 }
 
-exports.name = 're';
+exports.name = 'j';
 exports.factory = factory;
