+++ lib/function/non-real/k.js
@@ -31,7 +31,7 @@ function factory (type, config, load, typed) {
    *            A complex number or array with complex numbers
    * @return {number | BigNumber | Array | Matrix} The real part of x
    */
-  var re = typed('k', {
+  var k = typed('k', {
     'number | Complex': function (x) {
       return 0;
     },
@@ -41,18 +41,18 @@ function factory (type, config, load, typed) {
     },
 
     'Array | Matrix': function (x) {
-      return deepMap(x, re);
+      return deepMap(x, k);
     },
 
     'Quaternion': function (x) {
-      return x.k();
+      return x.k;
     }
   });
 
-  re.toTex = {1: '\\Re\\left\\lbrace${args[0]}\\right\\rbrace'};
+  k.toTex = {1: 'unfinished'};
 
-  return re;
+  return k;
 }
 
-exports.name = 're';
+exports.name = 'k';
 exports.factory = factory;
