+++ lib/function/non-real/re.js
@@ -48,7 +48,7 @@ function factory (type, config, load, typed) {
     },
 
     'Quaternion': function (x) {
-      return x.r();
+      return x.r;
     }
   });
